requirements.txt
fastapi[all]
passlib[bcrypt]
uvicorn[standard]
sqlalchemy
uuid
datetime
python-jose[cryptography]
pymysql
python-dotenv